loop through any thank yous that are not transcoded after 3 minutes a…

+++ app/models/thank_you.rb
@@ -105,8 +105,12 @@ class ThankYou < ApplicationRecord
   end
 
   def self.failed_upload(user_id)
-    thank_you = ThankYou.where(creator: user_id, transcoded: false).where("created_at < ?", DateTime.now - 3.minutes)[0]
-    thank_you.destroy
+    thank_yous = ThankYou.where(creator: user_id, transcoded: false).where("created_at < ?", DateTime.now - 3.minutes)[0]
+    if thank_yous[0]
+      thank_yous.map { |thank_you|
+        thank_you.destroy
+      }
+    end
   end
 
   private
